deletion restricted

+++ src/App.jsx
@@ -11,6 +11,8 @@ import Header from './components/Header';
 // 20220912223415
 // https://6315b6ef33e540a6d38296a9.mockapi.io/notepad-app
 
+const ID = nanoid()
+
 function App() {
 
   const randomColor1 = Math.floor(Math.random() * 16777215).toString(16);
@@ -18,7 +20,7 @@ function App() {
 
   const [notes, setNotes] = useState([
     {
-      id: nanoid(),
+      id: ID,
       color1: randomColor1,
       color2: randomColor2,
       text: 'this is the note pad app text',
@@ -47,7 +49,7 @@ function App() {
   const addNote = (text) => {
     const date = new Date();
     const newNote = {
-      id: nanoid(),
+      id: ID,
       color1: randomColor1,
       color2: randomColor2,
       text: text,
@@ -67,37 +69,38 @@ function App() {
   const deleteNote = (id) => {
     const newNotes = notes.filter((note) => note.id !== id);
     //delete operation
-    axios.delete(
-      `https://6315b6ef33e540a6d38296a9.mockapi.io/notepad-app/${id}`,
-      setNotes(newNotes)
-    );
+    if (id == ID)
+      axios.delete(
+        `https://6315b6ef33e540a6d38296a9.mockapi.io/notepad-app/${id}`,
+        setNotes(newNotes)
+      );
   };
 
   return (
     <>
-     <div className={`${darkMode && 'dark-mode'}`}>
-     <Header handleToggleDarkMode={setDarkMode} />
-    <div className="container">
-      <Search handleSearchNote={setSearchText} />
-      <CardsList
-        notes={notes.filter((note)=> note.text.toUpperCase().includes(searchText.toLocaleUpperCase()))}
-        handleAddNote={addNote}
-        handleDeleteNote={deleteNote}
-      />
-      <ToastContainer
-        position="top-center"
-        autoClose={2000}
-        hideProgressBar={false}
-        newestOnTop={false}
-        closeOnClick
-        rtl={false}
-        pauseOnFocusLoss
-        draggable
-        pauseOnHover
-      />
-      <ToastContainer />
-    </div>
-     </div>
+      <div className={`${darkMode && 'dark-mode'}`}>
+        <Header handleToggleDarkMode={setDarkMode} />
+        <div className="container">
+          <Search handleSearchNote={setSearchText} />
+          <CardsList
+            notes={notes.filter((note) => note.text.toUpperCase().includes(searchText.toLocaleUpperCase()))}
+            handleAddNote={addNote}
+            handleDeleteNote={deleteNote}
+          />
+          <ToastContainer
+            position="top-center"
+            autoClose={2000}
+            hideProgressBar={false}
+            newestOnTop={false}
+            closeOnClick
+            rtl={false}
+            pauseOnFocusLoss
+            draggable
+            pauseOnHover
+          />
+          <ToastContainer />
+        </div>
+      </div>
     </>
   );
 }
